+++ packages/kit/src/vite/utils.js
@@ -1,6 +1,19 @@
 import fs from 'fs';
 import path from 'path';
+import { loadConfigFromFile } from 'vite';
 import { get_runtime_path } from '../core/utils.js';
+
+/**
+ * @param {import('vite').ConfigEnv} config_env
+ * @return {Promise<import('vite').UserConfig>}
+ */
+export async function get_vite_config(config_env) {
+	const config = (await loadConfigFromFile(config_env))?.config;
+	if (!config) {
+		throw new Error('Could not load Vite config');
+	}
+	return config;
+}
 
 /**
  * @param {...import('vite').UserConfig} configs