+++ packages/kit/test/apps/amp/svelte.config.js
@@ -1,5 +1,3 @@
-import * as path from 'path';
-
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
@@ -8,24 +6,7 @@ const config = {
 			hydrate: false
 		},
 
-		inlineStyleThreshold: Infinity,
-
-		vite: {
-			server: {
-				fs: {
-					allow: [path.resolve('../../../src')]
-				},
-				// TODO: required to support ipv6, remove on vite 3
-				// https://github.com/vitejs/vite/issues/7075
-				host: 'localhost',
-				// TODO: remove on vite 3
-				// https://github.com/vitejs/vite/pull/8778
-				watch: {
-					// perf, do not watch playwright output dir
-					ignored: ['**/test-results/**']
-				}
-			}
-		}
+		inlineStyleThreshold: Infinity
 	}
 };