+++ packages/kit/test/utils.js
@@ -205,3 +205,7 @@ export async function start_server(handler, start = 4000) {
 		}
 	};
 }
+
+export const plugin = process.env.CI
+	? (await import('../dist/vite.js')).sveltekit
+	: (await import('../src/vite/index.js')).sveltekit;